testing/web-platform/tests/atomics/window-and-any-worker/wait/no-spurious-wakeup/no-spurious-wakeup.html

<!doctype html>
<meta charset=utf-8>
<title></title>

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<body>
    <iframe id="myiframe" style="display: none" src="../empty.html" ></iframe>
</body>

<script>

/*---
description: >
  Test that Atomics.wait actually waits and does not spuriously wake
  up when the memory value is changed.
includes: [atomicsHelper.js]
---*/

window.addEventListener("load", function(e) {
  async_test(function(t) {
    let worker = new Worker('wait-on-zero-diff.js');
    worker.onmessage = this.step_func_done(function(e) {
      let ret = e.data[0];
      assert_true(Math.abs(ret) <= 1000);
    });
    let ia = new Int32Array(new SharedArrayBuffer(Int32Array.BYTES_PER_ELEMENT));
    worker.postMessage([ia.buffer]);
    setTimeout(function() {
      // Change the value, should not wake the agent.
      Atomics.store(ia, 0, 1);
      // Wake up worker.
      setTimeout(() => assert_equals(1, Atomics.wake(ia, 0)), 500);
    }, 500);
  }, "atomics-wait-no-spurious-wakeup");
  async_test(function(t) {
    let worker = new Worker('no-spurious-wakeup-worker.js');
    worker.onmessage = this.step_func_done(function(e) {
      let ret = e.data[0];
      assert_true(Math.abs(ret) <= 1000);
    });
    worker.postMessage([]);
  }, "atomics-wait-no-spurious-wakeup-from-worker");
  async_test(function(t) {
    let worker = new SharedWorker('no-spurious-wakeup-shared-worker.js');
    worker.port.start();
    worker.port.onmessage = this.step_func_done(function(e) {
      let ret = e.data[0];
      assert_true(Math.abs(ret) <= 1000);
    });
    worker.port.postMessage([]);
  }, "atomics-wait-no-spurious-wakeup-from-shared-worker");
  async_test(function(t) {
    let iframe = document.getElementById('myiframe');
    let contentWindow = iframe.contentWindow;
    let test = this;
    contentWindow.onmessage = function(e) {
      let worker = new Worker('wait-on-zero-diff.js');
      worker.onmessage = test.step_func_done(function(e) {
        let ret = e.data[0];
        assert_true(Math.abs(ret) <= 1000);
      });
      let ia = new Int32Array(new SharedArrayBuffer(Int32Array.BYTES_PER_ELEMENT));
      worker.postMessage([ia.buffer]);
      setTimeout(function() {
        // Change the value, should not wake the agent.
        Atomics.store(ia, 0, 1);
        // Wake up worker.
        setTimeout(() => assert_equals(1, Atomics.wake(ia, 0)), 500);
      }, 500);
    }
    contentWindow.postMessage([], '*');
  }, "atomics-wait-no-spurious-wakeup-from-iframe");
});

</script>
